+++ handlers/__init__.py
@@ -1,6 +1,6 @@
 """
 Handlers Package
-版本: rev2
+版本: rev2.3.1
 提供事件處理功能
 """